fix: complete UI-utils move for LottieUtils, fix stale Kover comment
Final whole-branch review flagged two leftovers from this session's earlier
work:

- LottieUtils.kt was the one file skipped by the *Utils.kt -> ui/utils move
  (bfa2624) because an early grep scoped to android.*/androidx.* imports
  missed that it extends the third-party LottieAnimationView (a View
  subclass). Moves it alongside the other 20 files and fixes the
  NoEntryView.kt/test import that referenced its old top-level location.

- lib/build.gradle.kts's "*SettingsRepositoryKt*" Kover exclusion comment
  still cited commonUtilsSettings's @get:NoCoverage annotation, which was
  deleted earlier this session (bfe5f44). applyDarkMode() is now the only
  top-level declaration in that file; its when-branches are fully tested by
  ApplyDarkModeTest, but the file class remains subject to the same
  Kover/JaCoCo branch-misattribution flakiness already documented for
  URLUtilsKt, so the exclusion is kept and the comment now describes that.
